Append `Ticks` to some component
